sdl3 mixer and image issues

+++ Engine/Src/Ancona/HAL/SDL/SpriteImpl.cpp
@@ -56,7 +56,7 @@ void priv::SpriteImpl::Draw(
         _sdlTexture,
         &src,
         &dest,
-        (double) transform.rotation(),
+        transform.rotation(),
         &sdlOrigin,
         (SDL_FlipMode) flip);
 }
